Refactored id/base_id mapping and logging results
